+++ src/translations/en.js
@@ -17,6 +17,7 @@ module.exports = {
   fieldInvalidForPostalCode: 'This postal code is not valid.',
   genericError: 'Something went wrong on our end.',
   hostedFieldsFailedTokenizationError: 'Please check your information and try again.',
+  hostedFieldsTokenizationCvvVerificationFailedError: 'Credit card verification failed. Please check your information and try again.',
   hostedFieldsTokenizationNetworkErrorError: 'Network error. Please try again.',
   hostedFieldsFieldsInvalidError: 'Please check your information and try again.',
   paypalAccountTokenizationFailedError: 'Something went wrong adding the PayPal account. Please try again.',